% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plop-package.R
\docType{package}
\name{plop-package}
\alias{plop-package}
\title{plop: A 'ggplot2' assistant}
\description{
Provides an interactive addin that helps users iterate on ggplot2 code. Given some code selection from your editor, the assistant will iteratively generate a plot, suggest changes to make it more informative, and then incorporate suggestions into a new plot.
}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co} (\href{https://orcid.org/0000-0001-5676-5107}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
