% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plop.R
\name{plop}
\alias{plop}
\title{Interactively create and improve ggplot2 plots}
\usage{
plop(context = rstudioapi::getActiveDocumentContext())
}
\arguments{
\item{context}{Information on the active RStudio document, typically from
\code{rstudioapi::getActiveDocumentContext()}. Optional.}
}
\value{
Launches a shiny app that does the following:
\enumerate{
\item Given some code selection, suggests bare-bones plotting boilerplate
\item Runs the plotting code
\item Makes 4 suggestions on how to improve the plot
\item Once a suggestion has been submitted, generates new plotting code.
\item Iterate on steps 2-4.
}

Once you're satisfied with the resulting code, click "Apply" to integrate
the lines back into your file.
}
\description{
Interactively create and improve ggplot2 plots
}
